backend/app/tasks/media.py
from io import BytesIO
from PIL import Image
from app.tasks.celery_app import celery_app
from app.integrations.minio import minio_client
from app.core.logging import logger

@celery_app.task(name="tasks.process_image")
def process_image(object_name: str, media_id: str):
    """
    1. Download original from MinIO
    2. Get real dimensions (width, height)
    3. Convert to WebP
    4. Create thumbnail 480px
    5. Upload WebP + thumbnail back to MinIO
    """
    # This is a stub for the actual implementation
    # In a real scenario, you'd use minio_client.client.get_object
    # and then PIL to process the image.
    logger.info("processing_image_start", object_name=object_name, media_id=media_id)
    
    # Placeholder for actual logic
    # ...
    
    logger.info("processing_image_complete", object_name=object_name)
