backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic-settings==2.2.1
sqlalchemy[asyncio]==2.0.28
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
structlog==24.1.0
tenacity==8.2.3
python-dotenv==1.0.1
alembic==1.13.1
python-multipart==0.0.9
ruff==0.3.2
mypy==1.9.0
types-python-jose==3.3.4.20240106
types-passlib==1.7.7.20240311
redis==5.0.1
httpx==0.27.0
aiomoney==0.2.1
celery==5.3.6
fastapi-mail==1.4.1
aiogram==3.4.1
jinja2==3.1.3
cryptography==42.0.5
bleach==6.1.0
slowapi==0.1.9
pytest==8.1.1
pytest-asyncio==0.23.5
pytest-cov==4.1.0
bandit==1.7.8
safety==3.1.0
miniopy-async==1.19.0
Pillow==10.3.0
meilisearch-python-sdk==3.0.0
python-slugify==8.0.4
premailer==3.10.0
